+++ uimaj-ep-textmarker-engine/src/main/java/org/apache/uima/textmarker/expression/number/SimpleNumberExpression.java
@@ -36,7 +36,7 @@ public class SimpleNumberExpression extends NumberExpression {
   }
 
   @Override
-  public double getFloatValue(TextMarkerStatement parent) {
+  public float getFloatValue(TextMarkerStatement parent) {
     return number.floatValue();
   }
